+++ src/client/java/fr/madu59/obe/client/compat/lootr/LootrCompat.java
@@ -6,9 +6,11 @@ import fr.madu59.obe.client.registry.Registry;
 import fr.madu59.obe.client.registry.SpecialModelGetter;
 import fr.madu59.obe.client.registry.SpecialModelGetter.SpecialModelProvider;
 import fr.madu59.obe.client.util.blockentity.ChestUtil;
+import fr.madu59.obe.client.util.blockentity.ShulkerBoxUtil;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.Sheets;
 import net.minecraft.client.renderer.blockentity.ChestRenderer;
+import net.minecraft.client.resources.model.sprite.SpriteId;
 import net.minecraft.resources.Identifier;
 import net.minecraft.world.level.block.ChestBlock;
 import net.minecraft.world.level.block.entity.BlockEntity;
@@ -17,17 +19,27 @@ import noobanidus.mods.lootr.common.api.LootrAPI;
 import noobanidus.mods.lootr.common.api.LootrRegistry;
 import noobanidus.mods.lootr.common.api.LootrTags;
 import noobanidus.mods.lootr.common.block.entity.LootrChestBlockEntity;
+import noobanidus.mods.lootr.common.block.entity.LootrShulkerBoxBlockEntity;
+import noobanidus.mods.lootr.common.client.block.LootrChestBlockRenderer;
+import noobanidus.mods.lootr.common.client.block.LootrShulkerBoxRenderer;
 
 public class LootrCompat {
 
-    public static final Identifier MATERIAL = LootrAPI.rl("entity/chest/normal");
-    public static final Identifier MATERIAL2 = LootrAPI.rl("entity/chest/normal_opened");
-    public static final Identifier MATERIAL3 = LootrAPI.rl("entity/chest/trapped");
-    public static final Identifier MATERIAL4 = LootrAPI.rl("entity/chest/trapped_opened");
+    public static final Identifier CHEST_MATERIAL = LootrChestBlockRenderer.MATERIAL.texture();
+    public static final Identifier CHEST_MATERIAL2 = LootrChestBlockRenderer.MATERIAL2.texture();
+    public static final Identifier CHEST_MATERIAL3 = LootrChestBlockRenderer.MATERIAL3.texture();
+    public static final Identifier CHEST_MATERIAL4 = LootrChestBlockRenderer.MATERIAL4.texture();
+
+    public static final Identifier SHULKER_BOX_MATERIAL = LootrShulkerBoxRenderer.MATERIAL.texture();
+    public static final Identifier SHULKER_BOX_MATERIAL2 = LootrShulkerBoxRenderer.MATERIAL2.texture();
 
     public static void init(){
         Registry.addBlockEntityTypeInGroup("chest", LootrRegistry.getChestBlockEntity(), LootrRegistry.getTrappedChestBlockEntity());
         SpecialModelGetter.register(LootrRegistry.getChestBlockEntity(), new SpecialModelProvider(ChestUtil::getChestModelLayerLocation, LootrCompat::getChestMaterial, LootrCompat::transformChest, LootrCompat::getChestCacheKey));
+        SpecialModelGetter.register(LootrRegistry.getTrappedChestBlockEntity(), new SpecialModelProvider(ChestUtil::getChestModelLayerLocation, LootrCompat::getChestMaterial, LootrCompat::transformChest, LootrCompat::getChestCacheKey));
+    
+        Registry.addBlockEntityTypeInGroup("shulker_box", LootrRegistry.getShulkerBoxBlockEntity());
+        SpecialModelGetter.register(LootrRegistry.getShulkerBoxBlockEntity(), new SpecialModelProvider(ShulkerBoxUtil::getShulkerBoxModelLayerLocation, LootrCompat::getShulkerBoxMaterial, ShulkerBoxUtil::transformShulkerBox, LootrCompat::getShulkerBoxCacheKey));
     }
 
     public static Identifier getChestMaterial(BlockState state, BlockEntity be) {
@@ -42,13 +54,24 @@ public class LootrCompat {
                 }
             }
             if (isOpened) {
-                return isTrapped ? MATERIAL4 : MATERIAL2;
+                return isTrapped ? CHEST_MATERIAL4 : CHEST_MATERIAL2;
             } else {
-                return isTrapped ? MATERIAL3 : MATERIAL;
+                return isTrapped ? CHEST_MATERIAL3 : CHEST_MATERIAL;
             }
         }
         return ChestUtil.getChestMaterial(state);
     }
+
+    public static Identifier getShulkerBoxMaterial(BlockState state, BlockEntity be) {
+        if(be instanceof LootrShulkerBoxBlockEntity lootrShulkerBoxBe){
+            if (LootrAPI.isVanillaTextures()) {
+                return Sheets.DEFAULT_SHULKER_TEXTURE_LOCATION.texture();
+            } else {
+                return Minecraft.getInstance().player != null && lootrShulkerBoxBe.hasClientOpened(Minecraft.getInstance().player.getUUID()) ? SHULKER_BOX_MATERIAL2 : SHULKER_BOX_MATERIAL;
+            }
+        }
+        return ShulkerBoxUtil.getShulkerBoxMaterial(state);
+   }
 
     public static void transformChest(BlockState state, BlockEntity be, PoseStack poseStack) {
         poseStack.pushPose();
@@ -58,5 +81,9 @@ public class LootrCompat {
 
     public static Object getChestCacheKey(BlockEntity be) {
         return be instanceof LootrChestBlockEntity lootrChestBe && Minecraft.getInstance().player != null && lootrChestBe.hasClientOpened(Minecraft.getInstance().player.getUUID());
+    }
+
+    public static Object getShulkerBoxCacheKey(BlockEntity be) {
+        return be instanceof LootrShulkerBoxBlockEntity lootrShulkerBoxBe && Minecraft.getInstance().player != null && lootrShulkerBoxBe.hasClientOpened(Minecraft.getInstance().player.getUUID());
     }
 }