+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # What's new:
-- Added compatibility with Lootr's trapped chests
+- Added compatibility with Lootr's trapped chests and shulker boxes
 - Improved chunk meshing time thanks to an improved caching system
 - Added optimization for some custom skulls (which in most cases aren't optimized due to their dynamic texture online) e.g. Hypixel Skyblock's Fairy Souls, this also allow resource packs to override their texture if they want to
 - Added Russian translation (thanks @Redls07 and @mpustovoi)